refactor (#198)
